r431 Cleaned out Norm2 and L2Norm of templated commented out code and marked Norm2 as deprecated.
In LevelTree_Impl, put the logger(VERBOSE, ) where there was committed out cout.

Removed commented out code from Mesh.

In MeshManipulator, removed the commented out code and added a comment in which 
revision the commented out code is still available. Commented out cout's are 
converted to either logger(VERBOSE, ) or logger(DEBUG, ).

In MeshManipulator, added a todo for writing it to a stream.