r499 ----------Big constructor cleanup commit----------
In BaseBasisFunction, moved the using to the top of the public part.

In BasisFunctionSet, removed "using BaseBasisFunctionT = BaseBasisFunction" explicitly deleted the copy constructor and 
default constructor with "= delete" and added a comment that the default constructor and copy constructor are deleted. 
Also made the other constructor explicit.

In Edge, added a default constructor = delete, added "= delete" at the copy constructor and added a comment that the 
default constructor and copy constructor are deleted.

In Element, removed "using BasisFunctionSetT = Base::BasisFunctionSet" and "using ElementDataT = Base::ElementData". 
Spread the "parameter"-constructor and copy constructor over multiple lines and explained why the mass matrix is not 
copied.

Implemented a copy constructor in ElementData. Note that userData_ is not copied deeply, since it is an abstract base 
class and we typically want to copy the child implemented by the user. If deep copy is necessary, we need something 
like a virtual copy pattern, which defines a clone() function in every class related to UserElementData.

Added "= delete" to the copy constructor for Face, since we don't want someone copying a Face by accident (if adjacent 
elements are the same, then the Face already exists and there is no need for another, if adjacent elements are 
different, the copy is not really a copy).

Made the constructor in FaceCacheData explicit.

In FaceData, made a default constructor = delete and copy constructor = delete.

In FaceFactory, added "= delete" after the copy constructor and deleted the copy constructor, since it is a singleton 
and should not be copied.