r512 Did some minor cleanup in files that are committed but not mentioned below.
Deleted some default constructor from the files instead of doing "MyClass() = delete;" in the header files.

Started with the copy constructor of Mesh. When a Mesh is copied, we want actually a new Mesh instead of just copying 
the pointers to the old elements. 

Explicitly deleted the copy constructor of Element, since either:
- We create exactly the same element with the same faces etc., so the faces etc. are coupled to both the old and new 
element, and we can change e.g. the basis functions on one element and not the other. This is very prone to bugs.
- We create a new element with new faces, edges and nodes. But then we end up with too many faces, edges and nodes, 
since they are used for multiple elements. 

Instead of a copy constructor of Element, we want a function that copies an Element without the Face, Edge and Node 
vectors and makes a deep copy of the other data. Checked the copy constructor of ElementGeometry, this one already made 
a deep copy. ElementData does not require a deep copy. 

This commit is not a nice finished block of changes, but since I already changed so much, this commit is there to make 
a backup on the svn.