r766 New feature: (in progress)
GlobalSundialsVector is introduced in this commit. It is however still being build, hence in progress.

The vector will be the interface between the hpGEM memory structure and the Sundials N_Vector.
It will represent a vector of the global system.
Think of a RHS vector of all elements, or the solution vector of the system.

This vector will be able to do, in addition to DG, also conforming FEM.