r932 Use getters in preprocessors for template fields
The gcc did not accept the template-casting as done in the previous
version of the preprocessor. This casting is needed as an Element
contains information about not only its d-dimensional shape, but also
its d-1 dimensional faces, etc. till 0-dimensional points. Thus to 
access the lower dimensional data one needs to cast ElementShape.

With the new version we cast already in ElementShape, where both gcc
and (hopefully) clang will accept it.