r938 Fix the preprocessor to work with gcc
The implementation of the iterators causes compiler errors with the hybrid-
Iterator that is used as part of the stackVector. Therefore we replaced the
usage of stackVector by standard std::vectors.

The std::vector is far less efficient than the stackVector, as the latter has
some inbuild space for storing vectors with limited length. This was used as the
maximum number of nodes for an element is rather limited, hence having many
short std::vectors results in many allocations and many pointer traversals,
which degrade the performace.

If this performance is needed in the future, we might want to fix the
stackvector, or go for a similar approach where small vectors can be inlined.