r943 Fix (Div)DGMaxEigenvalue for parallel computation
The number of boundary faces may differt for different nodes in a
parallel computation. The number of MatAssemblyBegin/End needs to be the
same on all nodes of the cluster. This change ensures that the latter
is the case, by computing the maximum number of faces on the boundary of
a node. This is an upper bound to the number of MatAssemblyBegin/End
pairs.