r978 Added application extractsurface which can be run as an additional preprocessor step
The application expects a preprocessed mesh and a hyperplane orthogonal to one of the cardinal directions as input. The applications has been designed for situations where this hyperplane is the boundary of the domain. It will take all faces on the hyperplane and output them as elements on the output mesh file, where partitioning, the coordinates that were not fixed, and connectivity between copied elements will be preserved.

This commit also fixes a bug where empty meshes no longer crash the preprocessor.