r981 Add Brezzi style fluxes for DivDGMax
This code implements and Brezzi style fluxes for DivDGMax. it was based
on the work of Sarmany who implemented the Brezzi style fluxes for a
previous version of DGMax. The two additional stabilization parameters
in DivDGMax were replaced by Brezzi style fluxes using the paper from
Arnold et al. (Unified Analysis of Discontinuous Galerkin Methods for
Elliptic Problems) for the p/q variable. The normal jump one was done
by myself, but in the same style as the others.

Looking at the results so far the tangential jump flux and flux for p
are working as intended. The normal flux for u, does not seem to be
necessary and actually degrades the solution of the eigenvalue problem.