r989 Remove default basis construction
Remove that MeshManipulator automatically constructs monomial basis
functions. Most of this is directly removing checks for the existence
of basis functions (now we can have an empty 
CollectionOfBasisFunctionSetS), or setting the basis functions in the
tests. Most other codes already set their basis functions and thus
needed no modification.

A larger difference is in Element.cpp, where we needed to fix that the
integration rules for faces were never updated after construction. So
each face would have a quadrature rule suitable for order 0 after
construction. When configuring basis functions the quadrature rule for
the element would be adjusted, but those of the faces would remain at
this very low order one. Now the update on the element will check its
faces and update their quadrature rules as necessary.