r992 Remove BoundaryType
Boundary type was used to specify the types of boundaries used in
generating the mesh. As this mesh generation is now delegated to the
PreProcessor, it is no use to have this paramter left in
MeshManipulator, hence this removal. The two legacy mesh construction
methods were changed to keep their old behaviour