r1012 Rewrite MPI communication of GlobalIndexing
Recent testing showed that the communicatePushPullElements did not work
correctly when using multiple machines. The complexity discovered in
debugging it combined with the inefficiency of sending single messages
resulted in the decision to rewrite the method.

The purpose of the method it is to communicate the globalIndices of 
elements that are ghost elements for other processs. For elements this
is relatively simple as all the elements are in the push/pullElements
lists of the mesh. The global indices of the other geometric objects 
(faces, edges and nodes) are more complicated. As explained in the
comments it can happen that there is no push-pull pair which
can communicate the global index.

Therefore we split the communication in two rounds. The first round
communicates all the information for the objects that are owned by the
push processor. The second round communicates the information which is
only available after the first round. Care is taken to minimize the
amount of information that is transfered in the second round.

In addition to fixing these problems we also optimized the transmission
to only send one message per round to each processor (in addition to two
barriers). Thereby hopefully reducing transmission overhead.