r1023 Fix bug in GlobalIndexing communication
012345678901234567890123456789012345678901234567890123456789012345678912
The deduplication of sending indices for faces, edges and nodes (last 
only in intention, because of a secondary bug) was not correct.
Specifically when the owning element of a face, edge or node was not a
ghost element of the targetProcessor, then the global index information
for that geometric object would not be send for any element. There were
two solutions to this problem that I came up with:
 - Use the same idea as getOwningElement() == element, but instead of
   sending it for the first element, only send if for the first
   element that is a ghost element of the target processor. Using only
   geomtrical information is doable, but not easy and probably not
   efficient. While using the pushList is also not really attractive.
 - Store for each geometric object that has been send a tag, and check
   before adding the information if the tag was already added.

The latter option is far simpler and probably has good enough 
performance, hence that is the solution implemented in this commit.