r1026 Autodetect whether PETSc is real or complex
In the configuration of PETSc (for cmake) we could tell whether we want
a PETSc with real number (hpGEM_USE_PETSC) or complex numbers 
(hpGEM_USE_COMPLEX_PETSC). However, we did not check if the PETSc that
was found actually matches the requested type.

This comment changes that by using petscconf.h to detect the type of 
numbers of PETSc. Additionally I merged the configuration of PETSc into
one block and introduced the variable hpGEM_USE_ANY_PETSC (cmake), 
HPGEM_USE_ANY_PETSC (preprocessor) to replace constant or-ing of the
HPGEM_USE_PETSC and HPGEM_USE_COMPLEX_PETSC.