r1036 Fix nompi build on linux not finding dladdr
The linker complained about not being able to find dladdr when building
without any of the optional libraries (like PETSc, SLEPc and MPI). 
According to the man page of dladdr on linux one not only needs to 
include dlfcn.h but also link with -ldl. After adding this linker flag
it builds on both linux and OSX.

I did not test whether the failure to build without -ldl was also 
happening on OSX.