Compute eigenvalues for first k-space point in DGMax
The computation of a path through reciprocal space was limited to a path
with the origin as first point, for which eigenvalues were computed in
DGMax but not in DivDGMax. This commit removes the limitation of
requiring the origin as first point and the difference between
(Div)DGMax by always computing the eigenvalues for the first point. The
implementations for both discretizations have been modified to also
allow for a non zero first k-point.