Fix changing basis functions for mixed formulation
When using a scheme with multiple unknowns where at least one has a
different type of basis function, calling use***BasisFunctions(order),
the one without the unknown-parameter, would crash. The cause was that
the basis functions would be deleted at the start of the method, when
setting the basis function on each element through
setDefaultBasisFunctionSet it would do so unknown by unknown. When
setting the first unknown only the information of the first unknown
would be cleared, the out-dated information for the other unkowns would
remain. As the code for setting the basis function also updates the
quadrature rules, it tried to determine the maximum order of these old
basis functions. As these were already deleted the code did crash.

The fix here is to start by clearing all the information from old basis
functions. This is not ideal as it is also done again by
setDefaultBasisFunctionSet(position, unknown), but the easiest fix for
this probably uncommon situation.