Replace the sparsity estimates
The original code was undocumented with respect to its working and non
obvious in the way it worked. While easy to analyse for DG-type elements
it was hard to impossible to judge how correct it was for conforming
elements. The new code works much simpler, for each DoF it computes the
other DoFs it may share an element or face matrix with. The number of
these is the number of possible non-zero elements.