Rewrite the k-shifting in DGMaxEigenValue
The face matrices for the boundary faces need a shift e^(ikx) with k the
current k-vector and x the difference in coordianate between the left
and right face. This commit rewrites the implementation to not extract
the matices from the global matrix and incrementally shift them, but
instead to take the original assembled face matrices and shift them.
This has the advantage that only a single assembly step is needed for
the matrices. Moreover, this will make implementing shifts with
conforming finite elements much easier.