Keep a reference to the Mesh in GlobalIndexing
The Global Indexing is always with respect to a choice of basis
functions and a mesh. Therefore, it is logical to include a reference to
the mesh in the indexing. GlobalVector and GlobalMatrix did require both
Mesh and GlobalIndexing, which is now reduced to only needing the
indexing. Thereby preventing that the indexing is from a different mesh.