Add Hermitian DGMaxEigenvalue
The generalized eigenvalue problem Sx = lambda Mx of DGMax was always
reformulated to the non Hermitian eigenvalue problem M^{-1}Sx = lambda x.

In this commit we add the possibility to instead use the Chollesky
factorization LL^H = M to reformulate it as L^{-1}SL^{-H}y = lambda y,
with y = L^H x. This keeps the Hermitian structure of the matrix and
thus also the use of Hermitian eigenvalue solvers.