Remove old mesh generation/reading from kernel
The old (before the preprocessor) ways of reading and generating meshes
were accidentally readded in 12198e60. They were kept due to some bugs
causing issues with the same meshes from the preprocessor. As these bugs
all seem to have been fixed, we now again remove the generation of
square meshes in the kernel, just as the support for reading centaur
meshes directly.