Complete refactor of K-shifted boundary for DGMax
In DGMaxEigenvalue.cpp a complete implementation for the k-phase-shifted
boundary conditions for the problem was implemented. This took a
significant portion of the code and was very complex due to only minimal
abstraction. In this commit the k-phase-shifted boundary is moved into
two abstractions:
 - MatrixBlocks, a single or symmetricly placed pair of matrices in a
   global matrix that can be inserted into the global matrix.
 - KPhaseShift, the abstract implementation of a k-phase-shifted
   boundary condition for DG-stiffness matrices and the CG-DG projector
   matrix of DGMaxEigenvalue. This implementation is abstract in the
   sense that it is not tied to DGMaxEigenvalue, but could be used with
   any DG stiffness matrix.

With these two classes the implementation of the k-phase-shifted
boundary conditions in DGMaxEigenvalue becomes much simpler as we only
have to handle the specific details related to the various options
(shifts, hermitian, projector).