Make EVConvergence test more flexible in comparison
The output of the eigenvalue solver is determined by two factors:
 - The input matrix (discretization, mesh, parameters)
 - The implementation details of the eigenvalue solver
The input matrix determines that analytical eigenvalues, but which of
these are found and to what accuracy is determined by the implementation
of the eigenvalue solver. With different versions of PETSc/SLEPc the
implementation will change and thus result in slighly different
eigenvalues.

This was most prominent in DGMax, where eigenvalues may spiral towards
the zero eigenvalues resulting in different results for different
versions. An exact comparison with results from a previous run is
therefore not feasible as an integration test. Therefore, this commit
adds some flexbility in the comparison.