Refactor DGMaxEigenvalue SolverWorkspace
The SolverWorkspace was very intertwinde with the solve method of
DGMaxEigenvalue and complicated due to the three options (phase-shifts,
hermitian, projector). This commit reworks both the SolverWorkspace and
the solve method to break this mess.

The division is now that the solve() method of DGMaxEigenvalue should
run the high level algorithm of solving over k-points, but should not do
any numerical work itself. That is part of the methods called in
SolverWorkspace.

The handling of the three options is partially split of from
SolverWorkspace into ShiftWorkspace (optional phaseshifts) and
ProjectorWorkspace (optional Projector). The handling of Hermitian vs
non-Hermitian is left in the SolverWorkspace as this very much affects
the matrices, which are accessed by the Shift/ProjectorWorkspace