Make the KPhaseShift builders more flexible
There were two KPhaseShiftBuilders, each designed around the usecase of
DGMax (+Projector). One for a DG stiffness matrix where all DoFs in some
of the FaceMatrices should be given a phase shift. The other for the
projector (CGDG) where there is a element-matrix only global matrix,
but where the rows are of CG type and columns of DG. The continuity
requirements on the CG basis functions introduce the need for
phaseshifts there.

For standard (current) DivDGMax this is sufficient as both the DoFs for
the field and the Lagrange multiplier are of DG type. However, when
using CG basis functions for the multiplier we get a stiffness matrix
with both DG and CG DoFs. It was not possible to shift these with the
previous implementation and several changes had to be made:
 - For phase-shifting the DG-DG part the original
   FaceMatrixPhaseShiftBuilder was adapted so that it is possible to
   select only a subset of the unknowns for the actual phase shift.
 - For the DG-CG part this was also needed, but in addition the handling
   of the Hermitian part of the element matrix (DG rows, CG columns) is
   added.
 - There will be no CG-CG part, phase shifting this is still not
   supported due to the (expected) complexity in handling it.