Fix the labeling of subdomain boundary faces
The current check is to look if the same (by id) global nodes are both
side of the face. This identical to checking their coordinates but
independent of the dimension. This differs from the previous version
where no only the node coordinates were checked, but also their
ordering (which does not need to be the same on both sides).