Expand the correctly calculated owningElements
The preprocessor gives a layer of ghost elements around the owned part
of the mesh. This layer is such that any non owned element that shares
at least one node with an owned element is included. The previous check
on the safity of getOwningElement() in Face, Edge and Node was overly
restrictive in the context of CG basis functions. This commit widens the
checks.

For Node and Edge a basic check on the afforementioned principle is
used, is at least one element adjacent to the (end) Node(s) owned by the
current processor. For Face this is expanded by also considering faces
for which we know both sides. This is only possible as we know that
there are at most 2 elements to a face.