Work around for Petsc with zero inserts
The use of MAT_IGNORE_ZERO_ENTRIES has an oversight in combination with
MatSetValue(s). Calling MatSetValue(s) after the assembly with
ADD_VALUES correctly ignores zero values. For INSERT_VALUES there is no
check on whether the value is zero. So inserting a zero in a location
that was previously absent (as it was zero), results in an error.