Complete rewrite of the SparsityEstimator internals
The previous method worked and counted for each row DoF the number of
column DoFs with which it interacted. This is very efficient in storage
space, but the resulting code was horribly complex and hard to extend.

This replacement code takes a simpler route, all interactions between
row and column DoFs are counted simultaneously. This requires a bit more
storage (as all the interactions need to be stored), but this should
still be significantly less than for any sparse matrix. The counting
code goes over all element matrices and if applicable all face matrices.
For each of these the corresponding row and column DoFs will be
computed, resulting in the interactions.