Add field overlap computation to DGMax
This adds the posibility to compute the field-overlap integrals in
DGMax, these are the integrals of the form
integral_[unit cell] epsilon E_1 . E_2 dx.
We specifically compute the overlap between all the modes from the
previous k-vector and all the modes of the current wavevector. The idea
is that we may be able to use this information to compute the correct
connectivity of the modes.