Fix using direct solvers on CI
The use of direct solvers in PETSc is complicated. PETSc has its own
direct solver 'petsc'. However this only works on the seqaij matrix type
(the sequential one) not on the MPI version (mpiaij). The other solvers,
like MUMPS and UMFPACK, can be included, but that needs to be done at
build time. As we use two different CI platforms we get a complicated
state:

 - On Ubuntu the current version is petsc-3.7, and it has as
   dependencies all kinds of numerical solvers (MUMPS, UMFPACK, and
   more).
 - On MacOS we install everything by using homebrew. The homebrew
   version of petsc only has sparsesuite (UMFPACK) as dependency.
   Installing MUMPS would need to be done using the petsc installer.
   Which a) requires a lot of time and b) is not easily doable in
   homebrew.

So the first approach to use direct solvers in the CI was to just use
petsc's default internal solver. This requires switching the matrices
over from MPIAIJ to SEQAIJ when running on a single process, but that is
in fact an advantage, as that allows far more solvers for testing.

However, after doing this the DivDGMax selftest resulted in zero pivots
during the solve. Zero pivots that did not happen with either MUMPS of
UMFPACK. There did not seem to be a way to aleviate this problem with
the petsc solver. So for this specific test I used the options database
to switch it to using UMFPACK, the only solver available on both CI
platforms.

This is done via the options database, as that does not require direct
access to the lower solver configuration. However, the old petsc version
on Ubunutu presented a problem, as the function/key used to specify the
solver was renamed in petsc 3.9.