Periodic connecting in the Preprocessor (POC)
This is the first working version of taking a suitable non periodic
mesh and adding periodicity to it. Suitable in this case is that the
faces, edges and nodes are such that they could form a periodic
connection without hanging nodes.

The current POC is limited in that it only accepts unit hypercubes that
are aligned with the axes.