Refactor predefined structures in DGMax
For interesting computations each element needs information about the
material properties. This was/is done using some predefined structures
that are hard coded into DGMax. This approach is useful for some testing
and when working with structured meshes. However, it is quite limited
and impractical when working with externally generated meshes, as you
would have to code the material parameters yourself.

To aleviate this problem, this commit introduces the abstraction of a
StructureDescription, which maps each Element onto the material
parameters. The only implementation now is the Predefined version with
the historic structure definitions, but more will follow.