Add the ElementShape to its own parts
By adding the ElementShape to its boundary generating an array with all
its parts. This has the advantage that there is no distinction between
the boundary parts and the element, and thus that we can remove a whole
lot of conditional functions for the case entityDimension == dimension.

The disadvantage is that the ElementShape becomes non movable & non
copyable, as it contains a reference to itself. But moving and copying
them should not happen anyways due to using pointers to the individual
shapes.