Extract ReferenceGeometry lookup from ElementGeom
Moves the process of looking up the ReferenceGeometry for an element to
a factory class. This lookup is to find the ReferenceGeometry for an
element when the dimension and number of points is given. This lookup
is needed as the ReferenceGeometries do not have an unique identifier.
The extraction to a separate factory class allows much easier
modification and simple addition of the Lagrange geometries.