Automatic bidirectional links in Preprocessor mesh
MeshEntities and Elements are linked in a bidirectional fashion. This
was administrated as two separate single directional links, so a
MeshEntity could be linked to an Element but the Element not to the
MeshEntity (or vice versa). In this commit we modify the add/set methods
in Element so that they keep this link up to date.