Introduce merging Mesh parts after construction
This is primarily motivated by meshes with a mesh face with rotational
symmetry (e.g. 180 degrees). The initially constructed mesh can be
incorrect if:
 - There is a pair of connected nodes
 - Both of these nodes connect to a node inside the mesh (not on the
   surface).
The edges from both of the periodically connected nodes to the internal
node are then merged in their construction, which is incorrect (see also
the ascii art in MeshSource).

To prevent this from happening we need to construct the mesh without
identifying the periodically connected nodes. Then at later stage we
need to connect the nodes and merge some of the MeshEntities. For this
we only need to merge MeshEntities for which all coordinates (with
respect to one of the adjacent elements) are involved in the merging
process. This prevents the previously described unwanted merge, as the
coordinate of the internal node is not involved in such a merge.