Abstract the source of CoordinatTransformationData
The CoordinateTransformation (that actually transforms function values)
needs some information about the underlying coordinate transformation.
It was required that this came from a PhysicalElement, but in general
this does not need to be the case. This introduces a more general
interface so that different sources can be used as backing data for the
CoordinateTransformation.