Remove integrand scaling factor from CoordinateTransformation
The scaling should be the same for all unknowns. Hence placing it in the
CoordinateTransformation, which tells how the value of a unknown need to
be transformed, is illgocial. Such scaling should therefore be set on
the integral itself.

By doing this move we can remove the dependency of the
CoordinateTransformation on PhysicalElement/Face