Abstract the idea of vector<LazyCached<T>>
Having a vector<LazyCached<T>> has the disadvantage that each entry
needs its own computation. Resizing such a vector is therefore quite a
pain. This commit adds LazyVectorCached where all the entries in the
vector share the same computation, and this computation accepts as extra
argument the index in the vector.