Add div in 1D and curl in 1D&4D to basis function
This was done because the caching in physical face eventually
instantiated these. In the past this would only happen when actually
evaluating these specific divergence and curls.

The implementation for the divergence in 1D is straight forward, just
the derivative. The same implementation is used for the curl, as this
would allow generalizing curl-based codes to 1D in the same way that
they are generalized from 3D to 2D. The curl in 4D is not defined and
thus results in an error.