Refactor DGMaxDiscretization matrix construction
The primary step of this refactoring is to take the complicated matrix
assembly step and restructure it. The original implementation had a loop
over the elements/faces that did several things:
 - Run the integrator
 - Delegate the integrand to a function
 - Postprocess the resulting matrix/matrices for the mass matrix
 - Store the matrices as element/face matrix
 - [not modified] Compute element/face vectors

This refactoring unclutters this process to only two function calls:
 1. Call a function to compute the face/element matrices
 2. Call a function to post process the matrices

The first function does the integration for each type of matrix and
stores them in the element/face. The second one takes the matrix and the
and post processes them as required. I decided to keep them in this
fashion, as this would give the posibility of optimizing the
element/face integrals at some point.