Replace SmallVec/Mat impl by Eigen
This replaces most of the implementation of SmallVector and SmallMatrix
by using Eigen. The result is that SmallVector and SmallMatrix are just
wrappers around Eigen-matrices, which should prevent large overhead.

The wrapping classes should ensure that no aliasing happens, something
to look out for if one wants to remove the wrappers.