Add writing mutliple vtk output at once
To write multiple output streams before you had to call write() multiple
times, one for each stream. The problem with this is that expensive
computations could not be shared, e.g. writing complex valued vectors
requires computing the vector twice (real, imaginary). This becomes more
and more inefficient if more values need to be written.

This commit introduces an option to write multiple data streams at once.
For each point a shared type T is computed, which is then used to
compute the actual values to output for each stream.