Add complex valued diagonal material tensors
This adds the ability of having complex valued diagonal material
tensors. These are needed for the rectalinear PML. The choice for
supporting only diagonal tensors, is that there is no complex valued
small matrix (yet).

To keep the ElementInfos class simple, the tensors are always of
dimension 3. This also simplifies handling of the curl constant for 2D.

NOTE: The correct conjugation of the material constant in the Lagrange
mulitplier of DivDGMax has not been derived yet. The current
implementation is a guess based on symmetry. Specifically, all material
constants used in conjunction with test functions are conjugated.