Introduce abstraction for an analytic electric field
FieldPattern abstracts the concept of a known (analytic) Electric field,
with the assiciated derivatives. This allows abstracting the code
required for computing the source term of a harmonic problem.

Two of the sample problems are rewritten using the new
ExactFieldHarmonicProblem, a HarmonicProblem with known FieldPattern as
solution.

In the future this will allow scattering from such a known incident
field.