Fixes related to InterfaceReflectionField
The propagation at some point changed from x direction to y direction,
but the tests were not updated to reflect that. This changes the class
to have an arbitrary direction of interface and propagation, as long as
the normal and the propagation vector are parallel.

The latter condition is to have relatively simple reflection
coefficients.