Fix DGMaxDiscretization face integrals on MPI meshes
The main issue was that it crashed due to the way face matrices were
constructed. The FaceData implicutly assumed that the matrices are for
internal faces with equal number of right and left basis functions. For
faces on the subdomain boundary this is of course not the case.

The included fixes are:
 - Clarify MiddleSizeMatrix message (used to debug)
 - Include actual deprecation message (now that we have c++14)
 - Update DGMaxDiscretization to not compute the irrelevant face
   matrices, use correctly sized face matrices