Fix gmsh reading of multiple periodic boundaries
The gmsh reader incorrectly interpreted the number of periodic entities
number. As result it did not correctly read the case with >1 periodic
entities.

This commit fixes this issue, and migrates the handling of periodic
entities to the new style. That is, specify the periodic connections
rather than directly merging node coordinates.