Fix issues in Braggstack computation
Problem occured when fc & fd < 0. That way we would have a case with a
zero between [a, c] and a zero between [d,b], but instead the algorithm
would update a = c, removing the option to discover the zero between a
and c.

So to prevent this the fc < 0 and fd < 0 checks were moved up to before
updating the points, instead of after. This is slightly less efficient
because we check both points for being negative, but that is a very
minor inefficiency.