Doehler solver: Optimize projector
Previous version allocated a the solver and factored the matrix every
time the projector was called. This is wasteful as the matrix itself
does not change during the whole eigenvalue solve step. This changes it
so that the allocation is done beforehand and the factoring is done only
when the projector is used for the first time.